fix: add photo on User on ecommerce api

+++ src/main/java/com/list/ecommerce/repository/UsuarioRepository.java
@@ -7,8 +7,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
     Optional<Usuario> findByEmail(@NotBlank(message = "O campo não pode estar vazio") @Email(message = "Digite um email valido") String email);
 
-    Integer id(long id);
+    long id(long id);
 }
